chap3: Twelve days of christmas lyrics
PROBLEM: Print the lyrics to the Christmas carol "The Twelve Days of Christmas," taking advantage of the repetition in the song.

SOLUTION: Used fixed-size immutable arrays for the strings for both the gifts and the number words, plus nested loops to reverse the order of the gifts, and taking into account the mixed set nature of the for loop (half-inclusive). Punctuation is easier than it looks here. Aborted attempt to use u8 for the counter, due to lack of knowledge about customizing string indices with a particular integer type.

TEST:
cargo run & cargo run release. There is no user in put and this program always runs the same.